ed/za_voice中加入加载ogg api，d3d api， 及创建dsd的代码
这样的话dinput中无需再加载
