+++ asm/asm_dlgse.asm
@@ -15,6 +15,27 @@ dlgse_start:
 	mov     dword [ebx + tmp], eax
 	pop     eax
 
+	cmp     byte [ebx + order_dlgse], 0
+	je      short set_mute_over
+
+%ifdef za
+	mov     dword [esp + 0x0C], 0
+%else
+	mov     ecx, dword [ebx + addr_mute]
+	cmp     byte [ecx], 0
+	jne     short set_mute_over
+
+	mov     byte [ecx], 1
+	call    dword [ebx + to(jcs_dlgse)]
+	mov     ecx, dword [ebx + addr_mute]
+	mov     byte [ecx], 0
+	jmp     dlgse_stop
+%endif
+
+set_mute_over:
+	call    dword [ebx + to(jcs_dlgse)]
+
+dlgse_stop:
 	push    eax
 	push    ecx
 	push    edx
@@ -23,31 +44,7 @@ dlgse_start:
 	pop     edx
 	pop     ecx
 	pop     eax
-
-callend:
-	cmp     byte [ebx + order_dlgse], 0
-	je      short dlgse_return
-
-%ifdef za
-	mov     dword [esp + 0x0C], 0
-	call    dword [ebx + to(jcs_dlgse)]
-%else
-	mov     ecx, dword [ebx + addr_mute]
-	cmp     byte [ecx], 0
-	jne     short dlgse_return
-
-	mov     byte [ecx], 1
-	call    dword [ebx + to(jcs_dlgse)]
-	mov     ecx, dword [ebx + addr_mute]
-	mov     byte [ecx], 0
-%endif
-
-	push    dword [ebx + next(jcs_dlgse)]
-	mov     ebx, dword [ebx + tmp]
-	ret
-
-dlgse_return:
-	call    dword [ebx + to(jcs_dlgse)]
+	
 	push    dword [ebx + next(jcs_dlgse)]
 	mov     ebx, dword [ebx + tmp]
 	ret
