+++ solution/SoraVoice.sln
@@ -13,6 +13,7 @@ Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "ed_voice", "ed_voice\ed_voi
 EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "ed_voice_Test", "ed_voice_Test\ed_voice_Test.vcxproj", "{91D113FA-DDE9-46EF-8646-ACB606B82898}"
 	ProjectSection(ProjectDependencies) = postProject
+		{D8822198-F4C6-4DC4-AD4C-EA04A8277CCB} = {D8822198-F4C6-4DC4-AD4C-EA04A8277CCB}
 		{34C20CDE-F449-4ECB-9957-46E6E109956E} = {34C20CDE-F449-4ECB-9957-46E6E109956E}
 		{876751F5-E162-4093-B897-9914C9118034} = {876751F5-E162-4093-B897-9914C9118034}
 	EndProjectSection
