+++ solution/ed_voice/ed_voice.vcxproj
@@ -77,28 +77,29 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
+	<ClInclude Include="..\..\src\Mapping.h" />
     <ClInclude Include="..\..\src\Clock.h" />
     <ClInclude Include="..\..\src\Config.h" />
     <ClInclude Include="..\..\src\Draw.h" />
     <ClInclude Include="..\..\src\ed_voice.h" />
     <ClInclude Include="..\..\src\Hook.h" />
-    <ClInclude Include="..\..\src\Message.h" />
     <ClInclude Include="..\..\src\SoraVoice.h" />
     <ClInclude Include="..\..\src\InitParam.h" />
     <ClInclude Include="..\..\src\Log.h" />
-    <ClInclude Include="..\..\src\Mapping.h" />
     <ClInclude Include="..\..\src\rc\resource.h" />
+    <ClInclude Include="..\..\src\SoraVoiceImpl.h" />
     <ClInclude Include="..\..\src\SoundPlayer.h" />
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="..\..\src\Clock.cpp" />
     <ClCompile Include="..\..\src\Config.cpp" />
-    <ClCompile Include="..\..\src\Draw.cpp" />
+    <ClCompile Include="..\..\src\DrawImpl.cpp" />
     <ClCompile Include="..\..\src\ed_voice.cpp" />
     <ClCompile Include="..\..\src\Hook.cpp" />
+    <ClCompile Include="..\..\src\InitParam.cpp" />
     <ClCompile Include="..\..\src\Log.cpp" />
-    <ClCompile Include="..\..\src\SoraVoice.cpp" />
-    <ClCompile Include="..\..\src\SoundPlayer.cpp" />
+    <ClCompile Include="..\..\src\SoraVoiceImpl.cpp" />
+    <ClCompile Include="..\..\src\SoundPlayerImpl.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ResourceCompile Include="..\..\src\rc\ed_voice.rc" />
