+++ solution/ed_voice_Test/ed_voice_Test.vcxproj
@@ -52,7 +52,7 @@
     </ClCompile>
     <Link />
     <Link>
-      <AdditionalDependencies>ed_voice.lib;dxguid.lib;dsound.lib;libvorbis.lib;libvorbisfile.lib;dinput8.lib;%(AdditionalDependencies)</AdditionalDependencies>
+      <AdditionalDependencies>dxguid.lib;dsound.lib;libvorbis.lib;libvorbisfile.lib;dinput8.lib;%(AdditionalDependencies)</AdditionalDependencies>
     </Link>
   </ItemDefinitionGroup>
   <ItemDefinitionGroup Condition="'$(Configuration)|$(Platform)'=='Release|Win32'">
@@ -66,7 +66,7 @@
     <Link>
       <EnableCOMDATFolding>true</EnableCOMDATFolding>
       <OptimizeReferences>true</OptimizeReferences>
-      <AdditionalDependencies>ed_voice.lib;dxguid.lib;dsound.lib;libvorbis.lib;libvorbisfile.lib;dinput8.lib;%(AdditionalDependencies)</AdditionalDependencies>
+      <AdditionalDependencies>dxguid.lib;dsound.lib;libvorbis.lib;libvorbisfile.lib;dinput8.lib;%(AdditionalDependencies)</AdditionalDependencies>
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
