+++ solution/za_voice/za_voice.vcxproj
@@ -60,9 +60,7 @@
       <PreprocessorDefinitions>ZA;_SVDLL;%(PreprocessorDefinitions)</PreprocessorDefinitions>
     </ClCompile>
     <Link />
-    <Link>
-      <AdditionalDependencies>d3dx9d.lib;%(AdditionalDependencies)</AdditionalDependencies>
-    </Link>
+    <Link />
   </ItemDefinitionGroup>
   <ItemDefinitionGroup Condition="'$(Configuration)|$(Platform)'=='Release|Win32'">
     <ClCompile>
@@ -78,7 +76,6 @@
     <Link>
       <EnableCOMDATFolding>true</EnableCOMDATFolding>
       <OptimizeReferences>true</OptimizeReferences>
-      <AdditionalDependencies>d3dx9.lib;%(AdditionalDependencies)</AdditionalDependencies>
       <ImageHasSafeExceptionHandlers>false</ImageHasSafeExceptionHandlers>
     </Link>
   </ItemDefinitionGroup>
@@ -91,19 +88,20 @@
     <ClInclude Include="..\..\src\SoraVoice.h" />
     <ClInclude Include="..\..\src\InitParam.h" />
     <ClInclude Include="..\..\src\Log.h" />
-    <ClInclude Include="..\..\src\Mapping.h" />
     <ClInclude Include="..\..\src\rc\resource.h" />
+    <ClInclude Include="..\..\src\SoraVoiceImpl.h" />
     <ClInclude Include="..\..\src\SoundPlayer.h" />
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="..\..\src\Clock.cpp" />
     <ClCompile Include="..\..\src\Config.cpp" />
-    <ClCompile Include="..\..\src\Draw.cpp" />
+    <ClCompile Include="..\..\src\DrawImpl.cpp" />
     <ClCompile Include="..\..\src\ed_voice.cpp" />
     <ClCompile Include="..\..\src\Hook.cpp" />
+    <ClCompile Include="..\..\src\InitParam.cpp" />
     <ClCompile Include="..\..\src\Log.cpp" />
-    <ClCompile Include="..\..\src\SoraVoice.cpp" />
-    <ClCompile Include="..\..\src\SoundPlayer.cpp" />
+    <ClCompile Include="..\..\src\SoraVoiceImpl.cpp" />
+    <ClCompile Include="..\..\src\SoundPlayerImpl.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ResourceCompile Include="..\..\src\rc\ed_voice.rc" />
