+++ solution/za_voice/za_voice.vcxproj.filters
@@ -33,9 +33,6 @@
     <ClInclude Include="..\..\src\InitParam.h">
       <Filter>头文件</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\src\Message.h">
-      <Filter>头文件</Filter>
-    </ClInclude>
     <ClInclude Include="..\..\src\rc\resource.h">
       <Filter>头文件</Filter>
     </ClInclude>
@@ -48,12 +45,12 @@
     <ClInclude Include="..\..\src\Clock.h">
       <Filter>头文件</Filter>
     </ClInclude>
+    <ClInclude Include="..\..\src\SoraVoiceImpl.h">
+      <Filter>头文件</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="..\..\src\ed_voice.cpp">
-      <Filter>源文件</Filter>
-    </ClCompile>
-    <ClCompile Include="..\..\src\Draw.cpp">
       <Filter>源文件</Filter>
     </ClCompile>
     <ClCompile Include="..\..\src\Log.cpp">
@@ -62,16 +59,22 @@
     <ClCompile Include="..\..\src\Config.cpp">
       <Filter>源文件</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\src\SoraVoice.cpp">
-      <Filter>源文件</Filter>
-    </ClCompile>
     <ClCompile Include="..\..\src\Hook.cpp">
       <Filter>源文件</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\src\SoundPlayer.cpp">
+    <ClCompile Include="..\..\src\Clock.cpp">
       <Filter>源文件</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\src\Clock.cpp">
+    <ClCompile Include="..\..\src\SoraVoiceImpl.cpp">
+      <Filter>源文件</Filter>
+    </ClCompile>
+    <ClCompile Include="..\..\src\InitParam.cpp">
+      <Filter>源文件</Filter>
+    </ClCompile>
+    <ClCompile Include="..\..\src\DrawImpl.cpp">
+      <Filter>源文件</Filter>
+    </ClCompile>
+    <ClCompile Include="..\..\src\SoundPlayerImpl.cpp">
       <Filter>源文件</Filter>
     </ClCompile>
   </ItemGroup>
