+++ src/SoundPlayerImpl.cpp
